Move getCaretPositionDetails definition to setSelection module.
