+++ src/device/__tests__/splitAtSelection.ts
@@ -1,8 +1,7 @@
-import getCaretPositionDetails from '../getCaretPositionDetails'
 import splitAtSelection from '../splitAtSelection'
 import { setSelection } from '../../util/setSelection'
 
-it('split at selecion with nested nodes', () => {
+it('split at selection with nested nodes', () => {
   const thoughtValue = 'The <b>rise <i>and fall</i></b> of Nikola Tesla.'
   const dummyDiv = document.createElement('div')
 
@@ -12,12 +11,7 @@ it('split at selecion with nested nodes', () => {
   document.body.appendChild(dummyDiv)
 
   // The rise and| fall of Nikola Tesla. (Cursor should be after `and`)
-  // Note: getting required focus node and offset using relative offset
-  const caretPositionDetails = getCaretPositionDetails(dummyDiv, 12)
-
-  setSelection(caretPositionDetails?.focusNode || dummyDiv, {
-    offset: caretPositionDetails?.offset,
-  })
+  setSelection(dummyDiv, { offset: 12 })
 
   const selectionRange = document.getSelection()?.getRangeAt(0)
